refactor: remove unused components and styles, clean up project structure
