+++ src/App.tsx
@@ -2,13 +2,12 @@ import { useState, useEffect } from "react";
 import { Session } from "@supabase/supabase-js";
 import { Auth } from "@supabase/auth-ui-react";
 import { ThemeSupa } from "@supabase/auth-ui-shared";
-import { supaClient } from "./client";
-import Header from "./components/Header";
-import Footer from "./components/Footer";
+import { supaClient } from "./api/client";
 import CreateProfile from "./components/CreateProfile";
 
 export default function App() {
   const [session, setSession] = useState<Session | null>(null);
+
   useEffect(() => {
     supaClient.auth.getSession().then(({ data: { session } }) => {
       setSession(session);
@@ -24,17 +23,11 @@ export default function App() {
 
   if (!session) {
     return (
-      <>
-        <Header />
-        <main className="bg-primarylight flex justify-center m-auto w-screen h-[70vh]">
-          <Auth
-            supabaseClient={supaClient}
-            providers={["google"]}
-            appearance={{ theme: ThemeSupa }}
-          />
-        </main>
-        <Footer />
-      </>
+      <Auth
+        supabaseClient={supaClient}
+        providers={["google"]}
+        appearance={{ theme: ThemeSupa }}
+      />
     );
   } else {
     return <CreateProfile />;
