--- src/components/Footer.tsx
@@ -1,9 +1,0 @@
-const Footer = () => {
-    return(
-        <footer className="w-screen bg-primarydark h-[10vh]">
-            
-        </footer>
-    )
-}
-
-export default Footer
